숫자 표기단위 변경: Double -> BigDecimal

+++ fx-provider/clients/currency-layer/src/main/java/com/fastcampus/sr/fxprovider/clients/currency/layer/client/CurrencyLayerStubClient.java
@@ -3,6 +3,7 @@ package com.fastcampus.sr.fxprovider.clients.currency.layer.client;
 import com.fastcampus.sr.fxprovider.clients.currency.layer.dto.CurrencyLayerResponse;
 import com.fastcampus.sr.fxprovider.common.currency.Currency;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Arrays;
@@ -17,10 +18,10 @@ public class CurrencyLayerStubClient implements CurrencyLayerClient {
         currencyLayerResponse.setTimestamp(new Date().getTime());
         currencyLayerResponse.setSource(Currency.USD);
         currencyLayerResponse.setQuotes(Map.of(
-                "USDEUR", 0.987085,
-                "USDJPY", 138.346008,
-                "USDCNY", 6.751304,
-                "USDKRW", 1316.309794));
+                "USDEUR", BigDecimal.valueOf(0.987085),
+                "USDJPY", BigDecimal.valueOf(138.346008),
+                "USDCNY", BigDecimal.valueOf(6.751304),
+                "USDKRW", BigDecimal.valueOf(1316.309794)));
 
         return currencyLayerResponse;
     }
